chore: Update backend and frontend package dependencies
This commit updates the dependencies in the backend and frontend packages. Specifically, it updates the nodemon package to version 3.1.4 in both the package.json and package-lock.json files. This update ensures that the latest stable version of nodemon is being used.

The changes also include updating the start script in the frontend package.json file to use port 3000 instead of port 3001.

These updates were made to keep the project up to date with the latest dependencies and to ensure compatibility with other packages and libraries.

```